+++ x/plans/keeper/keeper.go
@@ -9,6 +9,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/lavanet/lava/common"
+	commontypes "github.com/lavanet/lava/common/types"
 	"github.com/lavanet/lava/x/plans/types"
 )
 
@@ -50,6 +51,16 @@ func NewKeeper(
 
 func (k Keeper) BeginBlock(ctx sdk.Context) {
 	k.plansFS.AdvanceBlock(ctx)
+}
+
+// Export all plans from the KVStore
+func (k Keeper) ExportPlans(ctx sdk.Context) []commontypes.RawMessage {
+	return k.plansFS.Export(ctx)
+}
+
+// Init all plans in the KVStore
+func (k Keeper) InitPlans(ctx sdk.Context, data []commontypes.RawMessage) {
+	k.plansFS.Init(ctx, data)
 }
 
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
