+++ x/plans/keeper/plan.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	commonTypes "github.com/lavanet/lava/common/types"
 	"github.com/lavanet/lava/utils"
 	epochstoragetypes "github.com/lavanet/lava/x/epochstorage/types"
 	"github.com/lavanet/lava/x/plans/types"
@@ -71,16 +70,6 @@ func (k Keeper) PutPlan(ctx sdk.Context, index string, block uint64) {
 // GetAllPlanIndices gets from the KVStore all the plans' indices
 func (k Keeper) GetAllPlanIndices(ctx sdk.Context) (val []string) {
 	return k.plansFS.GetAllEntryIndices(ctx)
-}
-
-// Export all plans from the KVStore
-func (k Keeper) ExportPlans(ctx sdk.Context) []commonTypes.RawMessage {
-	return k.plansFS.Export(ctx)
-}
-
-// Init all plans in the KVStore
-func (k Keeper) InitPlans(ctx sdk.Context, data []commonTypes.RawMessage) {
-	k.plansFS.Init(ctx, data)
 }
 
 func (k Keeper) ValidatePlanFields(ctx sdk.Context, planToAdd *types.Plan) error {
